changed back to neotree and reorganized directory structure
